Import really old stuff from Jeff's local commits
This project was originally on Bitbucket:

    https://bitbucket.org/SorenRasmussen/pyofp

The "jsquyres" directory was stuff that Jeff did in this tree as a
first part of generating Fortran MPI bindings for Open MPI.  I think
it worked up to the point of generating all the Fortran header files
(note: I think the MPI_SIZEOF debacle was after this, so it might not
do the MPI_SIZEOF stuff...?).  Regardless, it's ~2 years out of date,
so I don't know what he current status of it is.

But it might be a good starting point for someone to pick up and use
to generate all the OMPI Fortran bindings...?